+++ packages/core/src/lib/implementation/report-to-gql.ts
@@ -69,9 +69,7 @@ function auditToGQL(audit: AuditReport): PortalAudit {
     formattedValue: audit.displayValue,
     ...(audit.details && {
       details: {
-        ...(audit.details.issues && {
-          issues: audit.details.issues.map(issueToGQL),
-        }),
+        issues: audit.details.issues.map(issueToGQL),
       },
     }),
   };