+++ packages/utils/perf/score-report/optimized0.ts
@@ -1,11 +1,12 @@
-import { CategoryRef, GroupRef, Report } from '@code-pushup/models';
-import { ScoredReport } from '../../src';
 import {
-  EnrichedAuditReport,
-  EnrichedScoredGroup,
-} from '../../src/lib/reports/scoring';
+  AuditReport,
+  CategoryRef,
+  GroupRef,
+  Report,
+} from '@code-pushup/models';
+import { ScoredGroup, ScoredReport } from '../../src/lib/reports/types';
 
-function groupRefToScore(audits: EnrichedAuditReport[]) {
+function groupRefToScore(audits: AuditReport[]) {
   return (ref: GroupRef) => {
     const score = audits.find(audit => audit.slug === ref.slug)?.score;
     if (score == null) {
@@ -17,16 +18,11 @@ function groupRefToScore(audits: EnrichedAuditReport[]) {
   };
 }
 
-function categoryRefToScore(
-  audits: EnrichedAuditReport[],
-  groups: EnrichedScoredGroup[],
-) {
+function categoryRefToScore(audits: AuditReport[], groups: ScoredGroup[]) {
   return (ref: CategoryRef) => {
     switch (ref.type) {
       case 'audit':
-        const audit = audits.find(
-          a => a.slug === ref.slug && a.plugin === ref.plugin,
-        );
+        const audit = audits.find(a => a.slug === ref.slug);
         if (!audit) {
           throw new Error(
             `Category has invalid ref - audit with slug ${ref.slug} not found in ${ref.plugin} plugin`,
@@ -35,9 +31,7 @@ function categoryRefToScore(
         return audit.score;
 
       case 'group':
-        const group = groups.find(
-          g => g.slug === ref.slug && g.plugin === ref.plugin,
-        );
+        const group = groups.find(g => g.slug === ref.slug);
         if (!group) {
           throw new Error(
             `Category has invalid ref - group with slug ${ref.slug} not found in ${ref.plugin} plugin`,